rewrite and add clone to setup structure
