+++ bin/clone
@@ -1,0 +1,65 @@
+#!/usr/bin/env python3
+
+"""
+Copyright (C) 2015-2016 MushMush Foundation
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+"""
+
+import argparse
+import asyncio
+import os
+import sys
+from snare.utils import logger
+from snare.cloner import Cloner
+from snare.utils.snare_helpers import str_to_bool
+
+def main():
+    if os.getuid() != 0:
+        print('Clone has to be run as root!')
+        sys.exit(1)
+    if not os.path.exists('/opt/snare'):
+        os.mkdir('/opt/snare')
+    if not os.path.exists('/opt/snare/pages'):
+        os.mkdir('/opt/snare/pages')
+    loop = asyncio.get_event_loop()
+    parser = argparse.ArgumentParser()
+    parser.add_argument("--target", help="domain of the site to be cloned", required=True)
+    parser.add_argument("--max-depth", help="max depth of the cloning", required=False, default=sys.maxsize)
+    parser.add_argument("--log_path", help="path to the error log file")
+    parser.add_argument(
+        "--css-validate", help="set whether css validation is required", type=str_to_bool, default=None
+    )
+    args = parser.parse_args()
+    if args.log_path:
+        log_err = args.log_path + "clone.err"
+    else:
+        log_err = "/opt/snare/clone.err"
+    logger.Logger.create_clone_logger(log_err, __package__)
+    print("Error logs will be stored in {}\n".format(log_err))
+    try:
+        cloner = Cloner(args.target, int(args.max_depth), args.css_validate)
+        loop.run_until_complete(cloner.get_root_host())
+        loop.run_until_complete(cloner.run())
+    except KeyboardInterrupt:
+        pass
+
+
+if __name__ == '__main__':
+    print("""
+    ______ __      ______ _   ____________
+   / ____// /     / __  // | / / ____/ __ \\
+  / /    / /     / / / //  |/ / __/ / /_/ /
+ / /___ / /____ / /_/ // /|  / /___/ _, _/
+/_____//______//_____//_/ |_/_____/_/ |_|
+    
+    """)
+    main()
